requirements.txt
openai>=1.0.0
requests>=2.31.0
jellyfish>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pyyaml>=6.0.0
websockets>=12.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
